Move local participant join to base/participants
